feat: add /version endpoint for app metadata (#11)
