+++ internal/constants/constants.go
@@ -1,0 +1,6 @@
+package constants
+
+const (
+	AppName    = "Shift Planner"
+	AppVersion = "v0.1.0"
+)
